@@ -2,6 +2,7 @@
 using Microsoft.Extensions.DependencyInjection;
 using System.Reflection;
 using MediatR;
+using ShopDevelop.Application.Entities.User.Queries.GetProfile;
 
 namespace ShopDevelop.Application;
 
@@ -12,8 +13,8 @@ public static class DependencyInjection
     {
         services.AddMediatR(x =>
                 x.RegisterServicesFromAssemblies(
-                    typeof(GetMiniProductListQuery).Assembly));
-
+                    typeof(GetMiniProductListQuery).Assembly,
+                    typeof(GetUserProfileQuery).Assembly));
 
         services.AddMediatR(config => config.RegisterServicesFromAssembly(
             Assembly.GetExecutingAssembly()));
