@@ -9,7 +9,7 @@
         public uint Amount { get; set; } 
         public decimal OrderTotal { get; set; }
         public DateTime CreatedDate { get; set; }
-        public User User { get; set; }
+        public ApplicationUser User { get; set; }
         public Guid UserId { get; set; }
         public Product Product { get; set; }
     }
