@@ -5,7 +5,7 @@
         public Guid Id { get; set; }
         public Product Product { get; set; }
         public Guid ProductId { get; set; }
-        public User User { get; set; }
+        public ApplicationUser User { get; set; }
         public Guid UserId { get; set; }
         public DateTime CreatedDate { get; set; }
     }
