@@ -1,9 +1,9 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using ShopDevelop.Application.Entities.Product.Queries.GetMinimizedProducts;
 using ShopDevelop.Domain.Interfaces;
 using ShopDevelop.Persistence.Entities.Product.Queries.GetMinimizedProducts;
+using ShopDevelop.Persistence.Entities.User.Queries.GetProfile;
 
 namespace ShopDevelop.Persistence;
 
@@ -14,8 +14,9 @@ public static class DependencyInjection
         IConfiguration configuration)
     {
         services.AddMediatR(x =>
-                x.RegisterServicesFromAssemblies(
-                    typeof(GetMiniProductListHandler).Assembly));
+            x.RegisterServicesFromAssemblies(
+                typeof(GetMiniProductListHandler).Assembly,
+                typeof(GetUserProfileHandler).Assembly));
 
         var connectionString = configuration["DefaultConnection"];
         services.AddDbContext<ApplicationDbContext>(options =>
