@@ -1,6 +1,4 @@
-﻿using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 using ShopDevelop.Application.Services.Product;
 
 namespace ShopDevelop.WebApi.Controllers;
@@ -14,7 +12,6 @@ public class ProductController : BaseController
         this.productService = productService;
 
     [HttpGet]
-    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
     public async Task<IActionResult> Index(Guid? id)
     {
         if(id != null)
